ValidationError formats issues as JSON
