+++ src/main/utils/index.ts
@@ -1,6 +1,5 @@
 export * from './arrays';
 export * from './bitmasks';
-export * from './errors';
 export * from './lang';
 export * from './objects';
 export * from './shapes';
